sql database connection

+++ src/main/java/Model/Product.java
@@ -1,5 +1,5 @@
 package Model;
 
-public record Product(int id, String productName, String description, int price, String category) {
+public record Product(int id, String productName, String description, float price, String category) {
 
 }
